Update tutorials.socially.angular2.step_08.md
The `<form>` tag defaults to `display:block` and causes the login menu to appear in the next line, even though `.login-buttons` have `display:inline-block` style assigned.
At this point I'm becoming more and more inclined to believe that the whole tutorial is missing some "general style rules", which would explain while all the pages so far looked so ugly, and why the author expects the login menu to appear on the right - perhaps someone forgot to mention earlier in the tutorial that I was supposed to copy&paste some styles, or install materials design, or bootstrap or something?
One of the things I very much liked about the original, official, Metor tutorial (the one that made me build a simple TODO-app), was that it loooked pretty from step 1. This might seem naive of me, but such details as aesthetics, play some subconscious role in decision making, and I've still haven't made my mind about whether or not to use angular2-meteor in my next project.