Some clarifications to Step 8
I'm not what is the decorator the author means: the word "decorator" is mentioned only once in this Step, and it is not given any name.
I'm also not sure what are the two services the next statement talks about - my wild guess is that these are services with names matching the annotations mentioned in the third sentence. 
I'm not sure how to fix this paragraph, as I'm new to these all concepts. I would suggest to make the tutorial more authoritative and simply tell me what to do, and not ask questions to the reader like "How can you simplify life here?" -- the reader like me, has not enough knowledge to answer it, and presenting three possible choices (the decorator, two services, and two annotations) is beyond my ability to judge. Even more so, when they are not explicitly named and/or linked to the documentation :)
Just give me "the best practice" answer :)
It may sound harsh, but for me, the primary reason to use Meteor at all, is that it promises to remove (some of) the burden of making architectural decisions from me.